requirements.txt
Flask
requests
rich
readchar
sqlalchemy
pyyaml
pytest
